Firefox receiver fix
ChatGPT command:
When I test the receiver on firefox, I get an uncaught ReferenceError: senderId is not defined. This should not happen and the log output should show those types of errors as well

Secondary command:
Show the entire fix for this last fix. And for any future fixes, output the whole file instead of parts of the file. This makes it easier for me to update my code