Add receiver alias (name) in list
ChatGPT command:
Instead of the list showing an id, make it possible for the receiver to type an alias for themself. Display the alias in the list. If the receiver hasn't typed an alias, then it's ok to use the id. Update the list in real time, when a receiver is typing their alias. Also, only show receivers in the list. A sender should not be shown.